Resolves low rolls during late hours
`srand(time(NULL));` is needed in functions where they are declared in
roll code. That means all of the code, such as `rs_d20` to `rs_ww.cpp`
needed this fix.

I'm sure the fix is a bit of a hack and won't completely resolve this
issue, but until someone other than myself or the developer comes up
with a way to fully resolve this issue, this will probably continue to
persist but be much harder to detect as a result.