Fixes rolls being extremely lower than normal during late hours
`srand(time(NULL));` is needed to correctly resolve the extremely low
rolls during late hours. Azander informed me of this issue when I
provided the fix for `rs_act.cpp`.

This should resolve #2, but it will continue to persist, as this is more
of a hack than an effective solution. Still, until Azander or myself
come up with a better solution, this is the best way to resolve it.
It'll be that much more harder to detect, but it's still there.